Do not fail if shmem does not exist during unlinking
A race condition between two shmem finalization may occur if
one of them is made from a helper process that only opens the
shared memory to finalize other process. In that case both proceses
may think that they are the last one and both will try to delete
the shared memory. Ignore error in that case.